requirements.txt
discord.py
python-dotenv
openai
google-generativeai
deepl
yt-dlp